requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2023.7.22
charset-normalizer==3.3.0
defusedxml==0.7.1
Django==4.2.5
django-debug-toolbar==4.2.0
django-filter==23.3
django-modelcluster==6.1
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
django-webpack-loader==2.0.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.4
l18n==2021.3
openpyxl==3.1.2
Pillow==10.0.1
pillow-heif==0.13.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tzdata==2023.3
urllib3==2.0.6
wagtail==5.1.2
wagtailmedia==0.14.4
webencodings==0.5.1
Willow==1.6.2
